python-manager/requirements.txt
# Required for spell_grammar_checker
language_tool_python
# Required for spell_grammar_checker
pyspellchecker
pdf2docx
PyMuPDF
minio
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
python-docx==0.8.11
nltk==3.9.2
# Required for identity_detector (reductor)
presidio-analyzer

